src/requirements.txt
aiohttp
asn1tools
asyncio_mqtt
click
gps
PyYAML
python-json-logger
sslkeylog